Move some String+Helpers to WooFoundation
