+++ WooCommerce/Classes/Extensions/String+Helpers.swift
@@ -5,47 +5,6 @@ import UIKit
 /// String: Helper Methods
 ///
 extension String {
-
-    /// Helper method to provide the singular or plural (formatted) version of a
-    /// string based on a count.
-    ///
-    /// - Parameters:
-    ///   - count: Number of 'things' in the string
-    ///   - singular: Singular version of localized string — used if `count` is 1
-    ///   - plural: Plural version of localized string — used if `count` is greater than 1
-    /// - Returns: Singular or plural version of string based on `count` param
-    ///
-    /// NOTE: String params _must_ include `%ld` placeholder (count will be placed there).
-    ///
-    static func pluralize(_ count: Int, singular: String, plural: String) -> String {
-        if count == 1 {
-            return String.localizedStringWithFormat(singular, count)
-        } else {
-            return String.localizedStringWithFormat(plural, count)
-        }
-    }
-
-    /// Helper method to provide the singular or plural (formatted) version of a
-    /// string based on a count.
-    ///
-    /// - Parameters:
-    ///   - count: Number of 'things' in the string
-    ///   - singular: Singular version of localized string — used if `count` is 1
-    ///   - plural: Plural version of localized string — used if `count` is greater than 1
-    /// - Returns: Singular or plural version of string based on `count` param
-    ///
-    /// NOTE: String params _must_ include `%@` placeholder (count will be placed there).
-    ///
-    static func pluralize(_ count: Decimal, singular: String, plural: String) -> String {
-        let stringCount = NSDecimalNumber(decimal: count).stringValue
-
-        if count > 0 && count < 1 || count == 1 {
-            return String.localizedStringWithFormat(singular, stringCount)
-        } else {
-            return String.localizedStringWithFormat(plural, stringCount)
-        }
-    }
-
     /// Helper method to remove the last newline character in a given string.
     ///
     /// - Parameters:
@@ -63,11 +22,6 @@ extension String {
         }
 
         return newText
-    }
-
-    /// A Boolean value indicating whether a string has characters.
-    var isNotEmpty: Bool {
-        return !isEmpty
     }
 
     /// Get quotation marks from Locale
